requirements.txt
bcrypt==5.0.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
cryptography==46.0.6
docutils==0.22.4
filetype==1.2.0
idna==3.11
invoke==2.2.1
Kivy==2.3.1
Kivy-Garden==0.1.5
kivy-garden.mapview==1.0.6
mapview==1.0.6
mysql-connector-python==9.6.0
packaging==26.0
paramiko==2.12.0
pycparser==3.0
Pygments==2.20.0
PyNaCl==1.6.2
python-dotenv==1.2.2
requests==2.33.1
six==1.17.0
sshtunnel==0.4.0
urllib3==2.6.3
